added press images & update compose overrides

+++ images/worker/install-app.sh
@@ -6,7 +6,10 @@ APP=$1
 
 cd /home/frappe/frappe-bench
 
-rm -rf "apps/$APP/.git"
+# if $APP == "press" skip removing .git
+if [ "$APP" != "press" ]; then
+    rm -rf "apps/$APP/.git"
+fi
 
 env/bin/pip install -e "apps/$APP"
